Arrowhead refactor
